fix(cache): Fix development `events-processor` Redis cache DB (#611)
## Context

In development environment, the `lago-api` is using the Redis DB `3` for the cache store while the `events-processor` using the Redis DB `0`. This cause the `charge-usage` cache to never expire in development environment when the organization has Clickhouse event store enabled.

## Description

This simply updates the `.env.development.default` to use the same Redis DB.