chore(db): Add `pghero` as part of the dev services (#619)
## Context

We rely on [`pghero`](https://github.com/ankane/pghero/) on production. It can be useful to easily debug some performance issue, unused indexes, etc.

## Description

This adds `pghero` as a dev container.

In order to do so I had to update the `db` container to start PG with a specific config file. This file was copied from the actual container and updated to includes the necessary config for `pghero`.