chore(events-processor): Improve `events-processor` dev experience (#622)
## Context

The dev docker compose defines the dependencies of the `events-processor` but does not actually set the `condition`. Also the current `air` configuration doesn't allow for graceful shutdown.

## Description

This adds the `depends_on` conditions and also ensure that `air` will send an interrupt and wait 10s on shutdown before killing the app to allow for a graceful shutdown of the `events-processor`.